+++ astropy/table/table.py
@@ -110,6 +110,12 @@ class TableColumns(OrderedDict):
             raise IndexError('Illegal key or index value for {} object'
                              .format(self.__class__.__name__))
 
+    def __setitem__(self, item, value):
+        if item in self:
+            raise ValueError("Cannot replace column '{0}'.  Use Table.replace_column() instead."
+                             .format(item))
+        super(TableColumns, self).__setitem__(item, value)
+
     def __repr__(self):
         names = ("'{0}'".format(x) for x in six.iterkeys(self))
         return "<{1} names=({0})>".format(",".join(names), self.__class__.__name__)
@@ -1433,6 +1439,36 @@ class Table(object):
                 existing_names.add(new_name)
 
         self._init_from_cols(newcols)
+
+    def replace_column(self, name, col):
+        """
+        Replace column ``name`` with the new ``col`` object.
+
+        Parameters
+        ----------
+        name : str
+            Name of column to replace
+        col : column object (list, ndarray, Column, etc)
+            New column object to replace the existing column
+
+        Examples
+        --------
+        Replace column 'a' with a float version of itself::
+
+            >>> t = Table([[1, 2, 3], [0.1, 0.2, 0.3]], names=('a', 'b'))
+            >>> float_a = t['a'].astype(float)
+            >>> t.replace_column('a', float_a)
+        """
+        if name not in self.colnames:
+            raise ValueError('column name {0} is not in the table'.format(name))
+
+        if self[name].indices:
+            raise ValueError('cannot replace a table index column')
+
+        t = self.__class__([col], names=[name])
+        cols = OrderedDict(self.columns)
+        cols[name] = t[name]
+        self._init_from_cols(cols.values())
 
     def remove_row(self, index):
         """